AI-IMAGE/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.10.6
pydantic-settings==2.8.1
jinja2==3.1.2
pyjwt==2.8.0
openai>=1.12.0
boto3==1.29.0
aiohttp==3.9.1
aiofiles==23.2.1
certifi==2025.1.31 
